+++ dev/ci/dist_files.sh
@@ -32,19 +32,6 @@ for file in $distFile
     fi
   done
 
-echo "Dist dir will be in: "$path
-echo "------------------"
-export distDir="upgrade views/css views/js views/img"
-for dir in $distDir
-  do
-    echo -n "Copy $dir "
-    if [ -d "$modules" ]; then
-      cp -vr $path/${dir}/. $PWD/${modules}/${dir}
-    else
-      cp -vr $path/${dir}/. $PWD/${dir}
-    fi
-  done
-
 echo -n "Copy include.less "
 if [ -d "$modules" ]; then
   cp -v $path/include.less $PWD/${modules}/dev/css/less/include.less