+++ src/models/classes/paymentMethod/AmexPaymentMethod.php
@@ -55,7 +55,7 @@ class AmexPaymentMethod extends PaymentMethod
     // todo: add coverage to this method
     public function getPaymentTab()
     {
-        $payment_tab = parent::getPaymentTab();
+        $payment_tab = $this->getDefaultPaymentTab();
 
         if (empty($payment_tab)) {
             return $payment_tab;