+++ tests/models/classes/paymentMethod/OneyPaymentMethod/getPaymentTabTest.php
@@ -7,112 +7,52 @@ namespace PayPlug\tests\models\classes\paymentMethod\OneyPaymentMethod;
  * @group class
  * @group payment_method_class
  * @group oney_payment_method_class
+ * @group debug
  */
 class getPaymentTabTest extends BaseOneyPaymentMethod
 {
     public $expected_tab;
     public $oney;
+    public $oney_schedule;
+    public $oney_context;
 
     public function setUp()
     {
         parent::setUp();
-
-        $this->class->set('name', 'oney');
-        $this->validate_adapter->shouldReceive('validate')
-            ->andReturnUsing(function ($method, $object) {
-                return (bool) $object;
-            });
-        $this->configuration->shouldReceive('getValue')
-            ->with('currencies')
-            ->andReturn('EUR');
-        $this->helpers['amount']->shouldReceive([
-            'validateAmount' => [
+        $this->oney_schedule = '2';
+        $this->oney_context = 'oney_context';
+        $this->class->shouldReceive([
+            'isValidOneyCartQty' => [
                 'result' => true,
             ],
-            'convertAmount' => 4242,
+            'isValidOneyAddresses' => [
+                'result' => true,
+            ],
+            'isValidOneyAmount' => [
+                'result' => true,
+            ],
+            'getOneyPaymentContext' => $this->oney_context,
         ]);
+
+        $this->configuration->shouldReceive('getValue')
+            ->with('PS_TAX')
+            ->andReturn(42);
+        $this->tools_adapter->shouldReceive('tool')
+            ->withArgs(['getValue', 'payplugOney_type'])
+            ->andReturn(2);
+        $this->tools_adapter->shouldReceive('tool')
+            ->withArgs(['getValue', 'io'])
+            ->andReturn(1);
         $this->helpers['cookies']->shouldReceive([
             'setPaymentErrorsCookie' => true,
-        ]);
-        $config_class = \Mockery::mock('ConfigClass');
-        $config_class->shouldReceive([
-            'getIsoCodeByCountryId' => 'fr',
-            'formatPhoneNumber' => '0612345678',
-            'getIsoFromLanguageCode' => 'fr',
-        ]);
-        $this->dependencies->configClass = $config_class;
-        $this->tools_adapter->shouldReceive('tool')
-            ->andReturnUsing(function ($method, $arg) {
-                switch ($method) {
-                    case 'getShopDomainSsl':
-                        return true;
-
-                    default:
-                        if ('payplugOney_type' == $arg) {
-                            return '2';
-                        }
-
-                        return '';
-                }
-            });
-
-        $this->expected_tab = [
-            'amount' => 4242,
-            'currency' => 'EUR',
-            'notification_url' => 'link',
-            'force_3ds' => false,
-            'hosted_payment' => [
-                'return_url' => 'link',
-                'cancel_url' => 'link',
-            ],
-            'metadata' => [
-                'ID Client' => 1,
-                'ID Cart' => 1,
-                'Website' => true,
-            ],
-            'allow_save_card' => false,
-            'shipping' => [
-                'title' => null,
-                'first_name' => 'Ipsum',
-                'last_name' => 'Lorem',
-                'company_name' => 'Payplug',
-                'email' => 'customer@payplug.com',
-                'landline_phone_number' => '0612345678',
-                'mobile_phone_number' => '0612345678',
-                'address1' => '1 rue de l\'avenue',
-                'address2' => null,
-                'postcode' => '75000',
-                'city' => 'Paris',
-                'country' => 'fr',
-                'language' => 'fr',
-                'delivery_type' => 'BILLING',
-            ],
-            'billing' => [
-                'title' => null,
-                'first_name' => 'Ipsum',
-                'last_name' => 'Lorem',
-                'company_name' => 'Payplug',
-                'email' => 'customer@payplug.com',
-                'landline_phone_number' => '0612345678',
-                'mobile_phone_number' => '0612345678',
-                'address1' => '1 rue de l\'avenue',
-                'address2' => null,
-                'postcode' => '75000',
-                'city' => 'Paris',
-                'country' => 'fr',
-                'language' => 'fr',
-            ],
-        ];
-
-        $this->oney = \Mockery::mock('OneyOldRepository');
-        $this->plugin->shouldReceive([
-            'getOney' => $this->oney,
         ]);
     }
 
     public function testWhenParentMethodReturnEmptyArray()
     {
-        $this->class->set('name', '');
+        $this->class->shouldReceive([
+            'getDefaultPaymentTab' => [],
+        ]);
         $this->assertSame(
             [],
             $this->class->getPaymentTab()
@@ -122,10 +62,7 @@ class getPaymentTabTest extends BaseOneyPaymentMethod
     public function testWhenCurrentCartIsntElligible()
     {
         $this->class->shouldReceive([
-            'getOneyPriceLimit' => [
-                'min' => 100,
-                'max' => 3000,
-            ],
+            'getDefaultPaymentTab' => $this->default_payment_tab,
         ]);
 
         $this->validators['payment']->shouldReceive([
@@ -135,21 +72,6 @@ class getPaymentTabTest extends BaseOneyPaymentMethod
             ],
         ]);
 
-        $this->cart_adapter->shouldReceive([
-            'isGuestCartByCartId' => false,
-        ]);
-        $this->cart_adapter->shouldReceive([
-            'nbProducts' => 1001,
-        ]);
-        $this->configuration_adapter->shouldReceive('get')
-            ->with('PS_TAX')
-            ->andReturn('1');
-        $this->configuration_adapter->shouldReceive('get')
-            ->with('PS_CURRENCY_DEFAULT')
-            ->andReturn('EUR');
-        $this->currency_adapter->shouldReceive([
-            'get' => 1,
-        ]);
         $this->assertSame(
             [],
             $this->class->getPaymentTab()
@@ -159,100 +81,48 @@ class getPaymentTabTest extends BaseOneyPaymentMethod
     public function testWhenRequiredFieldsIsNeeded()
     {
         $this->class->shouldReceive([
+            'getDefaultPaymentTab' => $this->default_payment_tab,
             'hasOneyRequiredFields' => true,
-            'getOneyPriceLimit' => [
-                'min' => 100,
-                'max' => 3000,
-            ],
         ]);
-
         $this->validators['payment']->shouldReceive([
             'isOneyElligible' => [
                 'result' => true,
             ],
         ]);
-        $this->helpers['phone']->shouldReceive([
-            'isPhoneNumber' => [
-                'result' => true,
-            ],
-        ]);
         $this->helpers['cookies']->shouldReceive([
-            'getPaymentDataCookie' => [],
+            'getPaymentDataCookie' => true,
         ]);
 
-        $this->cart_adapter->shouldReceive([
-            'isGuestCartByCartId' => false,
-        ]);
-        $this->cart_adapter->shouldReceive([
-            'nbProducts' => 1001,
-        ]);
-        $this->configuration_adapter->shouldReceive('get')
-            ->with('PS_TAX')
-            ->andReturn('1');
-        $this->configuration_adapter->shouldReceive('get')
-            ->with('PS_CURRENCY_DEFAULT')
-            ->andReturn('EUR');
-
-        $this->currency_adapter->shouldReceive([
-            'get' => 1,
-        ]);
         $this->assertSame(
             [],
             $this->class->getPaymentTab()
         );
     }
 
-    public function testWhenPaymentTabIsReturned()
+    public function testPaymentTabIsReturned()
     {
         $this->class->shouldReceive([
+            'getDefaultPaymentTab' => $this->default_payment_tab,
             'hasOneyRequiredFields' => false,
-            'getOneyPaymentContext' => [],
-            'getOneyPriceLimit' => [
-                'min' => 100,
-                'max' => 3000,
-            ],
         ]);
-
         $this->validators['payment']->shouldReceive([
             'isOneyElligible' => [
                 'result' => true,
             ],
         ]);
-        $this->helpers['phone']->shouldReceive([
-            'isPhoneNumber' => [
-                'result' => true,
-            ],
-        ]);
-        $oney_schedule = 2;
-        $this->tools_adapter->shouldReceive('tool')
-            ->with('payplugOney_type')
-            ->andReturn($oney_schedule);
 
-        $this->expected_tab['authorized_amount'] = $this->expected_tab['amount'];
-        $this->expected_tab['auto_capture'] = true;
-        $this->expected_tab['payment_method'] = 'oney_' . $oney_schedule;
-        $this->expected_tab['payment_context'] = [];
+        $expected_tab = $this->default_payment_tab;
+        $expected_tab['authorized_amount'] = $expected_tab['amount'];
+        $expected_tab['force_3ds'] = false;
+        $expected_tab['auto_capture'] = true;
+        $expected_tab['payment_method'] = 'oney_' . $this->oney_schedule;
+        $expected_tab['payment_context'] = $this->oney_context;
+        $expected_tab['hosted_payment']['return_url'] = 'link';
 
-        unset($this->expected_tab['allow_save_card'], $this->expected_tab['amount']);
+        unset($expected_tab['allow_save_card'], $expected_tab['amount']);
 
-        $this->cart_adapter->shouldReceive([
-            'isGuestCartByCartId' => false,
-        ]);
-        $this->cart_adapter->shouldReceive([
-            'nbProducts' => 1001,
-        ]);
-        $this->configuration_adapter->shouldReceive('get')
-            ->with('PS_CURRENCY_DEFAULT')
-            ->andReturn('EUR');
-        $this->currency_adapter->shouldReceive([
-            'get' => 1,
-        ]);
-
-        $this->configuration_adapter->shouldReceive('get')
-            ->with('PS_TAX')
-            ->andReturn('1');
         $this->assertSame(
-            $this->expected_tab,
+            $expected_tab,
             $this->class->getPaymentTab()
         );
     }