+++ changelog.md
@@ -1,6 +1,9 @@
 # Payplug module changelog
 
 ## Version 4.22.0
+- Feature :
+  - [PRE-3033](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/): Add new payment method: Wero & Bizum
+
 - Bugfix :
   - [SMP-3239](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2733): Integration complementary address from apple wallet